+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # next
 
-- **Chore**: Migrate from Flow to Typescript
+- **Feature**: Added support `treeCol` on `<Column>`.
+- **Chore**: Migrate from Flow to Typescript.
 - **Breaking**: Remove the `setRowHeight` property,`rowHeight` supports function values.
 
 # 3.8.4